docker/django/requirements.txt
Django==3.2
psycopg2>=2.9
channels
web3
py-solc-x
pillow
uvicorn
